added misc directory for analysis scripts; created script to compare pass-1 and pass-2 nSidis files

+++ analysis_scripts/BSA_rgc_fits_updated.cpp
@@ -183,7 +183,7 @@ int createHistogramForBin(TTree* data, const char* histName, int binIndex,
   int numEventsPosTarget = 0;
   int numEventsNegTarget = 0;
 
-  float currentVariable;
+  double currentVariable;
   data->SetBranchAddress(propertyNames[currentFits].c_str(), &currentVariable);
 
   // for (int entry = 0; entry < data->GetEntries(); ++entry) {
